+++ src/engine/movegen/perft.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <engine/defs.hpp>
+#include <engine/constants.hpp>
 
 class Board;
 
@@ -10,8 +10,8 @@ namespace perft
      * @brief perft Function
      *
      * @param depth
-     * @return long long
+     * @return int
      */
-    unsigned long long perft(Board &board, int depth);
+    int perft(Board &board, int depth);
 
 } // namespace perft
